# docs/svc/openapi.yaml
components:
    schemas:
        AddPackageRequest:
            properties:
                ecosystem:
                    type: string
                identifier:
                    type: string
            type: object
        AddPackageResponse:
            properties:
                already_exists:
                    type: boolean
                ecosystem:
                    type: string
                id:
                    format: int32
                    type: integer
                identifier:
                    type: string
            type: object
...
        ErrorResponse:
            properties:
                error:
                    type: string
            required:
                - error
            type: object
...
        ListVersionsResponse:
            properties:
                ecosystem:
                    type: string
                identifier:
                    type: string
                latest_version:
                    nullable: true
                    type: string
                versions:
                    items:
                        type: string
                    type: array
            type: object
...
        PackageListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/PackageListItem'
                    type: array
            type: object
...
        TaskListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/TaskListItem'
                    type: array
            type: object
        TriggerScanRequest:
            properties:
                version:
                    type: string
            type: object
        TriggerScanResponse:
            properties:
                ecosystem:
                    type: string
                identifier:
                    type: string
                status:
                    type: string
                task_id:
                    format: int32
                    type: integer
                version:
                    type: string
            type: object
...
info:
    description: API for querying package information and metadata
    title: Secure Package Registry API
    version: 1.0.0
openapi: 3.0.3
paths:
    /api/v1/admin/packages:
        get:
            description: List all watched packages for a given ecosystem.
            operationId: adminListPackages
            parameters:
                - description: Ecosystem to list packages for
                  in: query
                  name: ecosystem
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PackageListResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: List watched packages
            tags:
                - Admin
        post:
            description: Create a new package in the watch list and publish a package requested event.
            operationId: adminAddPackage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddPackageRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AddPackageResponse'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Add package to watch list
            tags:
                - Admin
    /api/v1/admin/packages/{ecosystem}/{identifier}/behavior:
        get:
            description: Return the pre-computed deduped behavioral analysis tree for a specific package version.
            operationId: adminGetBehavior
            parameters:
                - in: path
                  name: ecosystem
                  required: true
                  schema:
                    type: string
                - in: path
                  name: identifier
                  required: true
                  schema:
                    type: string
                - description: Package version to retrieve behavior for
                  in: query
                  name: version
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                format: binary
                                type: string
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Get behavioral analysis (deduped)
            tags:
                - Admin
    /api/v1/admin/packages/{ecosystem}/{identifier}/behavior/raw:
        get:
            description: Return the raw (non-deduped) behavioral analysis tree for a specific package version.
            operationId: adminGetBehaviorRaw
            parameters:
                - in: path
                  name: ecosystem
                  required: true
                  schema:
                    type: string
                - in: path
                  name: identifier
                  required: true
                  schema:
                    type: string
                - description: Package version to retrieve behavior for
                  in: query
                  name: version
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                format: binary
                                type: string
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Get behavioral analysis (raw)
            tags:
                - Admin
    /api/v1/admin/packages/{ecosystem}/{identifier}/scan:
        post:
            description: Trigger a behavioral analysis scan for a package version. Defaults to the latest version if none is specified.
            operationId: adminTriggerScan
            parameters:
                - in: path
                  name: ecosystem
                  required: true
                  schema:
                    type: string
                - in: path
                  name: identifier
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TriggerScanRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TriggerScanResponse'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Trigger behavioral analysis scan
            tags:
                - Admin
    /api/v1/admin/packages/{ecosystem}/{identifier}/versions:
        get:
            description: List all known versions for a package.
            operationId: adminListVersions
            parameters:
                - in: path
                  name: ecosystem
                  required: true
                  schema:
                    type: string
                - in: path
                  name: identifier
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListVersionsResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: List package versions
            tags:
                - Admin
    /api/v1/admin/tasks:
        get:
            description: List collection tasks with package context, optionally filtered by ecosystem.
            operationId: adminListTasks
            parameters:
                - description: Filter by ecosystem
                  in: query
                  name: ecosystem
                  schema:
                    type: string
                - description: Page number (default 1)
                  in: query
                  name: page
                  schema:
                    type: string
                - description: Page size (default 50, max 100)
                  in: query
                  name: page_size
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TaskListResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: List collection tasks
            tags:
                - Admin
    /api/v1/admin/tasks/{taskID}/artifact:
        get:
            description: Stream the behavioral analysis artifact (.jsonl) for a completed collection task.
            operationId: adminDownloadArtifact
            parameters:
                - in: path
                  name: taskID
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/x-ndjson:
                            schema:
                                format: binary
                                type: string
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Download task artifact
            tags:
                - Admin
...